Add assertions to TaskHistoryManager
There are several unhandled returns in the Task History file reading process.

Handling these returns gives more clarity on top of the I/O exceptions in the event of a file read error.

Thus, when I/O exceptions occur, the result of creation/deletion of the TaskHistory file will be asserted, narrowing down the causes for an I/O exception